Update file specification format
Pull the local path information out of the file specifications in the config.txt file. Have the local path information given in a separate config setting. Also do not hard-code the digital value when initializing of the PH sample buffer. Instead use the calibration information to set it to a specific PH value.